Separate main MyApp
